옷장추가 기능구현 완료 //옷장 테이블 familyId는 nut null이라서 //기능 테스하려면 테이블에 임의로 추가해야함

+++ src/main/java/net/softsociety/secretary/domain/Closet.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Data
 public class Closet {
 	private int closetNum;
-	private Long familyId;
+	private int familyId;
 	private String userId;
 	private String closetName;	
 }
